Change API paths to include /account/<account_id>/

+++ backend/tests/integration.py
@@ -66,8 +66,8 @@ def test_values_put_get_simple(app_client, authorization_header):
         Put a value, get a value.
     """
     data = [{'p': 'qqqq.wwww', 't': 1234567890.123456, 'v': 111.22}]
-    app_client.put('/api/values/', data=json.dumps(data), content_type='application/json')
-    r = app_client.get('/api/values/?p=qqqq.wwww&t0=1234567890&t1=1234567891&a=no', headers={'Authorization': authorization_header})
+    app_client.put('/api/accounts/1/values/', data=json.dumps(data), content_type='application/json')
+    r = app_client.get('/api/accounts/1/values/?p=qqqq.wwww&t0=1234567890&t1=1234567891&a=no', headers={'Authorization': authorization_header})
     expected = {
         'paths': {
             'qqqq.wwww': {
@@ -88,7 +88,7 @@ def test_values_put_get_noaggrparam_redirect(app_client, authorization_header):
     """
     t_from = 1330000000
     t_to = 1330000000 + 100*15*60 + 1
-    url = '/api/values/?p=aaaa.bbbb&t0={}&t1={}&max=10'.format(t_from, t_to)
+    url = '/api/accounts/1/values/?p=aaaa.bbbb&t0={}&t1={}&max=10'.format(t_from, t_to)
     r = app_client.get(url, headers={'Authorization': authorization_header})
 
     assert r.status_code == 301
@@ -110,9 +110,9 @@ def test_values_put_get_sort_limit(app_client, authorization_header):
         {'p': TEST_PATH, 't': 1330002000 + 560, 'v': 140},
         {'p': TEST_PATH, 't': 1330002000 + 760, 'v': 160},
     ]
-    app_client.put('/api/values/', data=json.dumps(data), content_type='application/json')
+    app_client.put('/api/accounts/1/values/', data=json.dumps(data), content_type='application/json')
 
-    url = '/api/values/?p={}&a=no&sort=desc&limit=2'.format(TEST_PATH)
+    url = '/api/accounts/1/values/?p={}&a=no&sort=desc&limit=2'.format(TEST_PATH)
     r = app_client.get(url, headers={'Authorization': authorization_header})
 
     expected = {
@@ -144,10 +144,10 @@ def test_values_put_few_get_aggr(app_client, authorization_header):
         {'p': TEST_PATH, 't': 1330002000 + 760, 'v': 160},
     ]
 
-    app_client.put('/api/values/', data=json.dumps(data), content_type='application/json')
+    app_client.put('/api/accounts/1/values/', data=json.dumps(data), content_type='application/json')
     t_from = 1330002000  # aggr level 0 - every 1 hour
     t_to = t_from + 1*3600
-    url = '/api/values/?p={}&t0={}&t1={}&a=0'.format(TEST_PATH, t_from, t_to)
+    url = '/api/accounts/1/values/?p={}&t0={}&t1={}&a=0'.format(TEST_PATH, t_from, t_to)
     r = app_client.get(url, headers={'Authorization': authorization_header})
 
     expected = {
@@ -183,9 +183,9 @@ def test_values_put_many_get_aggr(app_client, authorization_header):
         #     {'t': 1234567890.123456, 'v': 111.22}
         # ]}}
 
-        app_client.put('/api/values/', data=json.dumps(data), content_type='application/json')
+        app_client.put('/api/accounts/1/values/', data=json.dumps(data), content_type='application/json')
         #print(t_from, t_to)
-        url = '/api/values/?p={}&t0={}&t1={}&max=10&a=3'.format(TEST_PATH, t_from, t_to)
+        url = '/api/accounts/1/values/?p={}&t0={}&t1={}&max=10&a=3'.format(TEST_PATH, t_from, t_to)
         r = app_client.get(url, headers={'Authorization': authorization_header})
 
         expected = {
@@ -204,14 +204,14 @@ def test_values_put_many_get_aggr(app_client, authorization_header):
         #pprint(actual)
         assert expected == actual
     finally:
-        r = app_client.delete('/api/paths/?p={}'.format(TEST_PATH), headers={'Authorization': authorization_header})
+        r = app_client.delete('/api/accounts/1/paths/?p={}'.format(TEST_PATH), headers={'Authorization': authorization_header})
         assert r.status_code == 200
 
 def test_paths_delete_need_auth(app_client, authorization_header):
     TEST_PATH = 'test.values.put.many.get.aggr'
-    r = app_client.delete('/api/paths/?p={}'.format(TEST_PATH))
+    r = app_client.delete('/api/accounts/1/paths/?p={}'.format(TEST_PATH))
     assert r.status_code == 401
-    r = app_client.delete('/api/paths/?p={}'.format(TEST_PATH), headers={'Authorization': authorization_header})
+    r = app_client.delete('/api/accounts/1/paths/?p={}'.format(TEST_PATH), headers={'Authorization': authorization_header})
     assert r.status_code == 404  # because path is not found, of course
 
 def test_dashboards_widgets_post_get(app_client, authorization_header):
@@ -222,10 +222,10 @@ def test_dashboards_widgets_post_get(app_client, authorization_header):
     try:
         WIDGET = 'chart1'
         data = {'name': DASHBOARD + ' name', 'slug': DASHBOARD}
-        r = app_client.post('/api/dashboards/', data=json.dumps(data), content_type='application/json', headers={'Authorization': authorization_header})
+        r = app_client.post('/api/accounts/1/dashboards/', data=json.dumps(data), content_type='application/json', headers={'Authorization': authorization_header})
         assert r.status_code == 201
 
-        r = app_client.get('/api/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
+        r = app_client.get('/api/accounts/1/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
         expected = {
             'name': DASHBOARD + ' name',
             'slug': DASHBOARD,
@@ -247,11 +247,11 @@ def test_dashboards_widgets_post_get(app_client, authorization_header):
                 }
             ])
         }
-        r = app_client.post('/api/dashboards/{}/widgets/'.format(DASHBOARD), data=json.dumps(widget_post_data), content_type='application/json', headers={'Authorization': authorization_header})
+        r = app_client.post('/api/accounts/1/dashboards/{}/widgets/'.format(DASHBOARD), data=json.dumps(widget_post_data), content_type='application/json', headers={'Authorization': authorization_header})
         assert r.status_code == 201
         widget_id = json.loads(r.data.decode('utf-8'))['id']
 
-        r = app_client.get('/api/dashboards/{}/widgets/'.format(DASHBOARD), headers={'Authorization': authorization_header})
+        r = app_client.get('/api/accounts/1/dashboards/{}/widgets/'.format(DASHBOARD), headers={'Authorization': authorization_header})
         actual = json.loads(r.data.decode('utf-8'))
         widget_post_data['id'] = widget_id
         expected = {
@@ -275,11 +275,11 @@ def test_dashboards_widgets_post_get(app_client, authorization_header):
                 }
             ])
         }
-        r = app_client.put('/api/dashboards/{}/widgets/{}'.format(DASHBOARD, widget_id), data=json.dumps(widget_post_data), content_type='application/json', headers={'Authorization': authorization_header})
+        r = app_client.put('/api/accounts/1/dashboards/{}/widgets/{}'.format(DASHBOARD, widget_id), data=json.dumps(widget_post_data), content_type='application/json', headers={'Authorization': authorization_header})
         assert r.status_code == 204
 
         # make sure it was updated:
-        r = app_client.get('/api/dashboards/{}/widgets/'.format(DASHBOARD), headers={'Authorization': authorization_header})
+        r = app_client.get('/api/accounts/1/dashboards/{}/widgets/'.format(DASHBOARD), headers={'Authorization': authorization_header})
         actual = json.loads(r.data.decode('utf-8'))
         widget_post_data['id'] = widget_id
         expected = {
@@ -291,21 +291,21 @@ def test_dashboards_widgets_post_get(app_client, authorization_header):
         assert expected == actual
 
         # get a single widget:
-        r = app_client.get('/api/dashboards/{}/widgets/{}/'.format(DASHBOARD, widget_id), headers={'Authorization': authorization_header})
+        r = app_client.get('/api/accounts/1/dashboards/{}/widgets/{}/'.format(DASHBOARD, widget_id), headers={'Authorization': authorization_header})
         actual = json.loads(r.data.decode('utf-8'))
         expected = widget_post_data
         assert r.status_code == 200
         assert expected == actual
 
         # delete dashboard:
-        r = app_client.delete('/api/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
+        r = app_client.delete('/api/accounts/1/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
         assert r.status_code == 200
-        r = app_client.get('/api/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
+        r = app_client.get('/api/accounts/1/dashboards/{}'.format(DASHBOARD), headers={'Authorization': authorization_header})
         assert r.status_code == 404
 
     except:
         # if something went wrong, delete dashboard so the next run can succeed:
-        app_client.delete('/api/dashboards/{}'.format(DASHBOARD))
+        app_client.delete('/api/accounts/1/dashboards/{}'.format(DASHBOARD))
         raise
 
 @pytest.mark.skip(reason="no idea.")
@@ -315,9 +315,9 @@ def test_values_put_paths_get(app_client):
     """
     PATH = 'test.values.put.paths.get.aaaa.bbbb.cccc'
     data = [{'p': PATH, 't': 1234567890.123456, 'v': 111.22}]
-    app_client.put('/api/values/', data=json.dumps(data), content_type='application/json')
+    app_client.put('/api/accounts/1/values/', data=json.dumps(data), content_type='application/json')
 
-    r = app_client.get('/api/paths/?filter=test.values.put.paths.get.*')
+    r = app_client.get('/api/accounts/1/paths/?filter=test.values.put.paths.get.*')
     expected = {
         'paths': {
             'test.values.put.paths.get.*': [
@@ -329,24 +329,24 @@ def test_values_put_paths_get(app_client):
     actual = json.loads(r.data.decode('utf-8'))
     assert expected == actual
 
-    r = app_client.get('/api/paths/?filter=test.*&failover_trailing=false')
+    r = app_client.get('/api/accounts/1/paths/?filter=test.*&failover_trailing=false')
     actual = json.loads(r.data.decode('utf-8'))
     assert PATH in actual['paths']['test.*']
 
-    r = app_client.get('/api/paths/?filter=test.&failover_trailing=false')
+    r = app_client.get('/api/accounts/1/paths/?filter=test.&failover_trailing=false')
     assert r.status_code == 400
-    r = app_client.get('/api/paths/?filter=test.')  # same - failover_trailing=false is default option
+    r = app_client.get('/api/accounts/1/paths/?filter=test.')  # same - failover_trailing=false is default option
     assert r.status_code == 400
 
     for prefix in ['t', 'te', 'tes', 'test', 'test.', 'test.v']:
-        r = app_client.get('/api/paths/?filter={}&failover_trailing=true'.format(prefix))
+        r = app_client.get('/api/accounts/1/paths/?filter={}&failover_trailing=true'.format(prefix))
         actual = json.loads(r.data.decode('utf-8'))
         assert actual['paths'] == {}
         assert PATH in actual['paths_with_trailing'][prefix]
         for path in actual['paths_with_trailing'][prefix]:
            assert path[:len(prefix)] == prefix
 
-    r = app_client.get('/api/paths/?filter=test.*,test.values.*')
+    r = app_client.get('/api/accounts/1/paths/?filter=test.*,test.values.*')
     actual = json.loads(r.data.decode('utf-8'))
     assert PATH in actual['paths']['test.*']
     assert PATH in actual['paths']['test.values.*']
